requirements.txt
kafka-python==2.0.2
pydantic==2.8.2
prometheus-client==0.20.0
river==0.20.1
structlog==24.1.0
orjson==3.10.3
